requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
boto3==1.11.9             # via smart-open
boto==2.49.0              # via smart-open
botocore==1.14.9          # via boto3, s3transfer
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
docutils==0.15.2          # via botocore
gensim==3.8.1
idna==2.8                 # via requests
jmespath==0.9.4           # via boto3, botocore
kafka-python==1.4.7
nltk==3.4.5
numpy==1.18.1             # via gensim, scipy
python-dateutil==2.8.1    # via botocore
requests==2.22.0          # via smart-open
s3transfer==0.3.2         # via boto3
scipy==1.4.1              # via gensim
six==1.14.0               # via gensim, nltk, python-dateutil
smart-open==1.9.0         # via gensim
urllib3==1.25.8           # via botocore, requests
vadersentiment==3.2.1
